% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds3_lexer_parser.R
\name{pds3_lexer_parser}
\alias{pds3_lexer_parser}
\title{Access PDS3 Lexer & Parser}
\usage{
pds3_lexer_parser(debug = FALSE)
}
\arguments{
\item{debug}{(default: FALSE) Sets lexer/parser to debug mode for testing.}
}
\value{
A list with the following components:
  \describe{
    \item{lexer}{The rly lexer}
    \item{parser}{The rly parser}
  }
}
\description{
Provides access to the underlying lexer & parser to process PDS3 files
}
\details{
Typically you don't need to access this directly, but if you are processing
a large number of files you may want to cache this and pass it in to
`read_pds3` manually to prevent them from being re-created on each run.
}
\examples{
util <- pds3_lexer_parser()
}
